// frontend/src/app/noticias/interfaces/noticia.interface.ts
export interface Noticia {
    notas: Nota[];
}

export interface Nota {
    idnota:               number;
    idusuario:            number;
    titulo:               string;
    descripcion:          string;
    imagen:               null | string;
    fechahorapublicacion: Date;
    created_at:           Date;
    updated_at:           Date;
    usuario:              Usuario;
}

export interface Usuario {
    idusuario:      number;
    idpersonal:     number;
    nombreusuario:  string;
    password:       string;
    idtipopersonal: string;
    personal:       Personal;
}

export interface Personal {
    idpersonal:     number;
    apepaterno:     string;
    apematerno:     string;
    nombre:         string;
    direccion:      string;
    fechadeingreso: Date;
}

// Generated by https://quicktype.io

export interface Persona {
    idpersonal:     number;
    apepaterno:     string;
    apematerno:     string;
    nombre:         string;
    direccion:      string;
    fechadeingreso: string;
    user:           User;
}

export interface User {
    idusuario:      number;
    idpersonal:     number;
    nombreusuario:  string;
    password:       string;
    idtipopersonal: string;
}


// Generated by https://quicktype.io

export interface Comentario {
    comentarios: Comentario[];
}

export interface Comentario {
    idcomentario:         number;
    comentario:           string;
    idusuario_comentario: number;
    idnota:               number;
    created_at:           string;
    updated_at:           string;
    respuestas:           Respuesta[];
    usuario:              Usuario;
}

export interface Respuesta {
    id:            number;
    comentario_id: number;
    usuario_id:    number;
    respuesta:     string;
    created_at:    string;
    updated_at:    string;
    usuario:       Usuario;
}

export interface Usuario {
    idusuario:      number;
    idpersonal:     number;
    nombreusuario:  string;
    password:       string;
    idtipopersonal: string;
}


export interface EditUsuario {
    apepaterno: string; 
    apematerno: string; 
    nombre: string;
    direccion: string; 
    fechadeingreso: string; 
    nombreusuario: string; 
    idtipopersonal: number; 
}